% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_functions.R
\name{age_query}
\alias{age_query}
\title{Build a SQL expression for calculating patient age.}
\usage{
age_query(age_method)
}
\arguments{
\item{age_method}{Either "dob" (full date of birth) or "year_only".}
}
\value{
A SQL expression string.
}
\description{
Build a SQL expression for calculating patient age.
}
\keyword{internal}
