% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_truth_functions.R
\name{build_ground_truth_values_clause}
\alias{build_ground_truth_values_clause}
\title{Build a SQL VALUES clause from a resolved ground truth dataframe.}
\usage{
build_ground_truth_values_clause(resolved_gt, person_ids_batch, dialect)
}
\arguments{
\item{resolved_gt}{Dataframe as returned by \code{resolve_ground_truth_ids}.}

\item{person_ids_batch}{Vector of person_ids for this batch.}

\item{dialect}{SQL dialect, used for correct timestamp cast syntax.}
}
\value{
A character string like
  \code{(1, 100, CAST('2022-07-01 10:00:00' AS TIMESTAMP), ...)}.
  Suitable for substitution into \code{@ground_truth_values} in the SQL query.
}
\description{
Takes a resolved ground truth dataframe (as returned by
\code{resolve_ground_truth_ids}), filters to a batch of person_ids,
and returns a SQL VALUES string.
}
\note{
The VALUES clause is inlined into the SQL query, so its size scales
  with the number of ground truth rows per batch. With the default
  \code{batch_size = 10000}, a batch could produce a ~500KB SQL string.
  This works but is not elegant. If performance is a concern, reduce
  \code{batch_size} in \code{get_score_variables}.
}
