% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_ii_functions.R
\name{comorbidities}
\alias{comorbidities}
\title{Calculates comorbidity variable for APACHE II. Assumes no comorbidities if no data.}
\usage{
comorbidities(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables and units of measure.
Should be the output of the get_score_variables function with the 'severity score parameter set to APACHEII"}
}
\value{
A data frame with the physiology values converted to the default units of measure specified.
}
\description{
Calculates comorbidity variable for APACHE II. Assumes no comorbidities if no data.
}
