% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_ii_functions.R
\name{fix_implausible_values_apache_ii}
\alias{fix_implausible_values_apache_ii}
\title{Removes implausible values from variables used to calculate the APACHE II score.
Assumes the units of measure have been fixed using the fix_apache_ii_units function.}
\usage{
fix_implausible_values_apache_ii(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables and units of measure.
Should be the output of the get_score_variables function with the 'severity score parameter set to APACHE II"}
}
\value{
A data frame with implausible physiology values deleted.
}
\description{
Removes implausible values from variables used to calculate the APACHE II score.
Assumes the units of measure have been fixed using the fix_apache_ii_units function.
}
