% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa_functions.R
\name{fix_implausible_values_sofa}
\alias{fix_implausible_values_sofa}
\title{Removes implausible values from variables used to calculate the SOFA score.
Assumes the units of measure have been fixed using the fix_sofa_units function.}
\usage{
fix_implausible_values_sofa(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables and units of measure.
Should be the output of the get_score_variables function with the 'severity score parameter set to "SOFA"}
}
\value{
A data frame with implausible physiology values deleted.
}
\description{
Removes implausible values from variables used to calculate the SOFA score.
Assumes the units of measure have been fixed using the fix_sofa_units function.
}
