% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa_functions.R
\name{fix_sofa_units}
\alias{fix_sofa_units}
\title{Convert units of measure into a format for the SOFA score calculation}
\usage{
fix_sofa_units(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables and units of measure. Should be the
output of the get_score_variables function with the severity score parameter
set to "SOFA"}
}
\value{
A data frame with physiology values converted to
the default units of measure specified.
}
\description{
Assumes units of measure are encoded in OMOP using the UCUM source vocabulary
Throws a warning if the unit of measure is not recognised. Assumes the
default unit of measure if not available.
}
