% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_distributions.R
\name{get_physiology_variable_availability}
\alias{get_physiology_variable_availability}
\title{Get physiology variable availability.
Returns a dataframe of availability with variables arranged in alphabetical order.}
\usage{
get_physiology_variable_availability(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables and (optionally) APACHE II score.}
}
\value{
A dataframe listing variable names & availability % per patient day.
}
\description{
Get physiology variable availability.
Returns a dataframe of availability with variables arranged in alphabetical order.
}
\examples{
availability_dataframe <- get_physiology_variable_availability(data)
}
