% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_distributions.R
\name{get_physiology_variable_distributions}
\alias{get_physiology_variable_distributions}
\title{Create histogram of distributions of APACHE II physiology variables.}
\usage{
get_physiology_variable_distributions(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables. Should have had the units of measure function run as well.}
}
\value{
A dataframe listing variable names and % availability per patient day.
}
\description{
Create histogram of distributions of APACHE II physiology variables.
}
\examples{
get_physiology_variable_distributions(data)
}
