% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_ii_functions.R
\name{mean_arterial_pressure}
\alias{mean_arterial_pressure}
\title{Calculates mean arterial pressure if not already calculated.}
\usage{
mean_arterial_pressure(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables and units of measure.
Should be the output of the get_score_variables function with the 'severity score parameter set to APACHEII"}
}
\value{
A data frame with the physiology values converted to the default units of measure specified.
}
\description{
Calculates mean arterial pressure if not already calculated.
}
