% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa_functions.R
\name{mechanical_ventilation}
\alias{mechanical_ventilation}
\title{Calculates mechanical ventilation variable for SOFA
This should include both invasive and non-invasive ventilation.}
\usage{
mechanical_ventilation(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables and units of measure.
Should be the output of the get_score_variables function with the
severity score parameter set to "SOFA"}
}
\value{
A data frame with the physiology values converted
to the default units of measure specified.
}
\description{
Calculates mechanical ventilation variable for SOFA
This should include both invasive and non-invasive ventilation.
}
