% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_functions.R
\name{omop_connect}
\alias{omop_connect}
\title{Create a connection to an OMOP-compliant database.}
\usage{
omop_connect(driver, host, dbname, port, user, password)
}
\arguments{
\item{driver}{Driver name: "PostgreSQL" or a driver string for odbc.}

\item{host}{Host/server name.}

\item{dbname}{Database name.}

\item{port}{Port number.}

\item{user}{Username.}

\item{password}{Password.}
}
\value{
A DBI connection object.
}
\description{
Create a connection to an OMOP-compliant database.
}
