% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{postgres_connect}
\alias{postgres_connect}
\title{Create a connection object to uds.}
\usage{
postgres_connect(host, user, password)
}
\arguments{
\item{host}{postgres host}

\item{user}{postgres username}

\item{password}{postgres password}
}
\description{
Create a connection object to uds.
}
