% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_multiple_queries.R
\name{run_multiple_queries}
\alias{run_multiple_queries}
\title{Runs an SQL script containing multiple create , drop statements
There can only be one select statement, and this returns results from that.}
\usage{
run_multiple_queries(con, sql_script)
}
\arguments{
\item{sql_script}{The script to be executed}

\item{conn}{A connection object to a database}
}
\description{
Runs an SQL script containing multiple create , drop statements
There can only be one select statement, and this returns results from that.
}
