% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_functions.R
\name{string_search_expression}
\alias{string_search_expression}
\title{Internal function called in `get_score_variables`.
Builds a regex for string searches}
\usage{
string_search_expression(concepts, variable_names, table_name)
}
\description{
Internal function called in `get_score_variables`.
Builds a regex for string searches
}
