% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_ii_functions.R
\name{total_gcs}
\alias{total_gcs}
\title{Calculates total gcs if not already calculated.}
\usage{
total_gcs(data)
}
\arguments{
\item{data}{Dataframe containing physiology variables and units of measure. Should be the
output of the get_score_variables function with the severity score parameter
set to "APACHE II"}
}
\value{
A data frame with the physiology values converted
to the default units of measure specified.
}
\description{
Calculates total gcs if not already calculated.
}
