% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_functions.R
\name{translate_drug_join}
\alias{translate_drug_join}
\title{Translate a lateral join for the Drug table between PostgreSQL and SQL Server.}
\usage{
translate_drug_join(dialect)
}
\arguments{
\item{dialect}{"postgresql" or "sql server".}
}
\value{
A SQL fragment string.
}
\description{
Translate a lateral join for the Drug table between PostgreSQL and SQL Server.
}
\keyword{internal}
