% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_functions.R
\name{age_query}
\alias{age_query}
\title{Internal function called in `get_score_variables`.
Builds the age query}
\usage{
age_query(age_method, dialect)
}
\description{
Internal function called in `get_score_variables`.
Builds the age query
}
