% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_ii_functions.R
\name{calculate_apache_ii_score}
\alias{calculate_apache_ii_score}
\title{Calculates the APACHE II score.
Assumes the units of measure have been fixed using the fix_apache_ii_units function.
Missing data is handled using normal imputation by default. If 'none', there is no imputation.}
\usage{
calculate_apache_ii_score(data, imputation = "normal")
}
\arguments{
\item{data}{Dataframe containing physiology variables and units of measure.}

\item{imputation}{Should be the output of the get_score_variables function with the 'severity score parameter set to APACHE II"}
}
\value{
A data frame with a variable containing the apache II score calculated.
}
\description{
Calculates the APACHE II score.
Assumes the units of measure have been fixed using the fix_apache_ii_units function.
Missing data is handled using normal imputation by default. If 'none', there is no imputation.
}
