% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_score_variables}
\alias{get_score_variables}
\title{Queries the postgres database to get variables required for the severity score specified.
Fixes the postgres timezone issue.}
\usage{
get_score_variables(
  postgres_conn,
  schema,
  start_date,
  end_date,
  dataset_name,
  severity_score
)
}
\arguments{
\item{postgres_conn}{A connection object to the postgres database}

\item{schema}{The name of the schema you want to query.}

\item{start_date}{The date/datetime of the first admission. Needs to be in character format.}

\item{end_date}{As above, but for last date}

\item{dataset_name}{Describes which concept codes to select. Can only be CCHIC at the moment.}

\item{severity_score}{The name of the severity score to calculate. Only APACHE2 at the moment.}
}
\description{
Queries the postgres database to get variables required for the severity score specified.
Fixes the postgres timezone issue.
}
