% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_score_variables}
\alias{get_score_variables}
\title{Queries a database to get variables required for a specified severity score.
Assumes visit detail table contains ICU admission information. If not
available, uses visit_occurrence.}
\usage{
get_score_variables(
  conn,
  dialect,
  schema,
  start_date,
  end_date,
  min_day,
  max_day,
  concepts_file_path,
  severity_score
)
}
\arguments{
\item{conn}{A connection object to a database}

\item{dialect}{dialect A dialect supported by SQLRender}

\item{schema}{The name of the schema you want to query.}

\item{start_date}{The earliest ICU admission date/datetime. Needs to be in character format.}

\item{end_date}{As above, but for last date}

\item{min_day}{Days since ICU admission to get physiology data for, starting with 0.}

\item{max_day}{Days since ICU admission to get physiology data for.}

\item{severity_score}{The name of the severity score to calculate. Only APACHE II at the moment.}

\item{mapping_path}{Path to the custom *_concepts.tsv file containing score to OMOP mappings.
Should match the example_concepts.tsv file format.}
}
\description{
Queries a database to get variables required for a specified severity score.
Assumes visit detail table contains ICU admission information. If not
available, uses visit_occurrence.
}
