% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{omop_connect}
\alias{omop_connect}
\title{Create a connection object to on OMOP compliant database.}
\usage{
omop_connect(driver, host, dbname, port, user, password)
}
\arguments{
\item{driver}{driver eg. "SQL Server", "PostgreSQL", etc.}

\item{host}{host/server name}

\item{dbname}{name of database in server}

\item{port}{database port}

\item{user}{username}

\item{password}{password}
}
\description{
Create a connection object to on OMOP compliant database.
}
