% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_functions.R
\name{translate_drug_join}
\alias{translate_drug_join}
\title{Internal function called in `get_score_variables`.
Translates left lateral join between postgres and sql server.}
\usage{
translate_drug_join(dialect)
}
\description{
Internal function called in `get_score_variables`.
Translates left lateral join between postgres and sql server.
}
