% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{variables_query}
\alias{variables_query}
\title{Internal function called in `get_score_variables`.
Builds the 'variable required' query for each table}
\usage{
variables_query(
  concepts,
  table_name,
  concept_id_var_name,
  value_as_concept_id_var = "",
  table_id_var = ""
)
}
\description{
Internal function called in `get_score_variables`.
Builds the 'variable required' query for each table
}
