% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_functions.R
\name{variables_query}
\alias{variables_query}
\title{Internal function called in `get_score_variables`.
Builds the 'variable required' query for each table}
\usage{
variables_query(concepts, concept_id_var, table_id_var = "")
}
\description{
Internal function called in `get_score_variables`.
Builds the 'variable required' query for each table
}
