% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_functions.R
\name{window_query}
\alias{window_query}
\title{Internal function called in `get_score_variables`.
Builds the windowing query based on variable names in each table.}
\usage{
window_query(window_start_point, time_variable, date_variable, cadence)
}
\description{
Internal function called in `get_score_variables`.
Builds the windowing query based on variable names in each table.
}
